# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Patent Prosecution Notebook API
  version: 0.2.0
paths:
...
  /cases:
    post:
      summary: Create/import case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                jurisdiction: { type: string, enum: [CN, EU, US] }
                application_no: { type: string }
                publication_no: { type: string }
                family_strategy: { type: string }
      responses:
        "200": { description: Case created }
    get:
      summary: List cases
      responses:
        "200": { description: List cases }
  /cases/{id}/ingest:
    post:
      summary: Start ingestion (async)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                providers:
                  type: array
                  items: { type: string, enum: [cnipr, epo, uspto, dms] }
                prefer_official: { type: boolean, default: true }
                include_dms_fallback: { type: boolean, default: true }
                trigger_processing: { type: boolean, default: true }
      responses:
        "200": { description: Queued with task id and missing hints }
...